requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
asyncpg==0.29.0; python_full_version >= '3.8.0'
bcrypt==4.0.1; python_version >= '3.6'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
cryptography==42.0.5
databases==0.9.0; python_version >= '3.8'
dnspython==2.6.1; python_version >= '3.8'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.1.1
fastapi==0.110.0; python_version >= '3.8'
greenlet==3.0.3; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.4; python_version >= '3.8'
httpx==0.27.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
passlib[bcrypt]==1.7.4
pluggy==1.4.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic[email]==2.6.3; python_version >= '3.8'
pydantic-core==2.16.3; python_version >= '3.8'
pydantic-settings==2.2.1; python_version >= '3.8'
pytest==8.0.2; python_version >= '3.8'
pytest-html==4.1.1; python_version >= '3.8'
pytest-metadata==3.1.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-jose[cryptography]==3.3.0
python-multipart==0.0.9; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.28; python_version >= '3.7'
starlette==0.36.3; python_version >= '3.8'
typing-extensions==4.10.0; python_version >= '3.8'
uvicorn==0.27.1; python_version >= '3.8'
